feat: upgrade dependencies and update documentation
This commit upgrades several dependencies in the pom.xml, including Spring Boot, Spring Shell, Mapstruct, Spotbugs annotations, and Jacoco Maven Plugin. It also updates the Postgres Docker version in the README and adjusts application.yml. Additionally, the project description in pom.xml was updated to reflect the CLI nature of the application.